# docs/swagger.yaml
basePath: /api/v0
definitions:
...
  service.Amount:
    properties:
      amount:
        type: integer
      denom:
        type: string
    type: object
  service.Attribute:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  service.Event:
    properties:
      attributes:
        items:
          $ref: '#/definitions/service.Attribute'
        type: array
      type:
        type: string
    type: object
  service.Fee:
    properties:
      amount:
        items:
          $ref: '#/definitions/service.Amount'
        type: array
      gas:
        type: integer
    type: object
...
  service.Log:
    properties:
      events:
        items:
          $ref: '#/definitions/service.Event'
        type: array
      log:
        type: string
      msg_index:
        type: integer
      success:
        type: boolean
    type: object
  service.Message:
    properties:
      type:
        type: string
      value:
        type: object
    type: object
  service.MovieInfo:
    properties:
      country:
        type: string
      genre:
        type: string
      runningTime:
        type: integer
      score:
        type: number
      title:
        type: string
      year:
        type: integer
    type: object
...
  service.PriceInfo:
    properties:
      discount:
        type: integer
      grandTotal:
        type: integer
      subTotal:
        type: integer
      usedFungible:
        type: integer
      usedServiceToken:
        type: integer
    type: object
  service.PubKey:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  service.PurchaseInfo:
    properties:
      movieInfo:
        $ref: '#/definitions/service.MovieInfo'
        type: object
      priceInfo:
        $ref: '#/definitions/service.PriceInfo'
        type: object
      ticketInfo:
        $ref: '#/definitions/service.TicketInfo'
        type: object
    type: object
...
  service.Signature:
    properties:
      pubKey:
        $ref: '#/definitions/service.PubKey'
        type: object
      signature:
        type: string
    type: object
  service.TicketInfo:
    properties:
      date:
        type: string
      price:
        type: integer
      sit:
        type: string
      theater:
        type: string
    type: object
  service.Transaction:
    properties:
      code:
        type: integer
      gasUsed:
        type: integer
      gasWanted:
        type: integer
      height:
        type: integer
      index:
        type: integer
      logs:
        items:
          $ref: '#/definitions/service.Log'
        type: array
      raw_log:
        type: string
      timestamp:
        type: string
      tx:
        $ref: '#/definitions/service.Tx'
        type: object
      txhash:
        type: string
    type: object
  service.TransferRequestResult:
    properties:
      redirectUri:
        type: string
      requestSessionToken:
        type: string
    type: object
  service.Tx:
    properties:
      type:
        type: string
      value:
        $ref: '#/definitions/service.TxValue'
        type: object
    type: object
  service.TxValue:
    properties:
      fee:
        $ref: '#/definitions/service.Fee'
        type: object
      memo:
        type: string
      msg:
        items:
          $ref: '#/definitions/service.Message'
        type: array
      signatures:
        items:
          $ref: '#/definitions/service.Signature'
        type: array
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is sample dapp to provide trials of LBD service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Link Cinema API
  version: "0.1"
paths:
...
  /test/init:
    get:
      consumes:
      - application/json
      description: Transfer tokens to user
      produces:
      - application/json
      responses:
        "200":
          description: transaction hashes has executed
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Init asset for test user
      tags:
      - test
  /test/transaction:
    get:
      consumes:
      - application/json
      description: Retrieve a Transaction using its hash
      parameters:
      - description: Transaction hash used for searching
        in: query
        name: txhash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction with the provided hash
          schema:
            $ref: '#/definitions/service.Transaction'
      summary: Get a transaction
      tags:
      - test
  /ticket:
    get:
      consumes:
      - application/json
      description: Retrieve a purchase info about given movie ticket
      produces:
      - application/json
      responses:
        "200":
          description: Ticket info
          schema:
            $ref: '#/definitions/service.PurchaseInfo'
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a purchase info
      tags:
      - ticket
  /ticket/purchase:
    post:
      consumes:
      - application/json
      description: Request user to transfer token at LBW
      parameters:
      - description: Purchase info
        in: body
        name: purchase_info
        required: true
        schema:
          $ref: '#/definitions/service.PurchaseInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Session token and redirect url to transfer token
          schema:
            $ref: '#/definitions/service.TransferRequestResult'
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Request user to purchase
      tags:
      - ticket
  /ticket/purchase/commit/{baseCoinTransferToken}/{:movieTokenTransferToken}:
    post:
      consumes:
      - application/json
      description: Commit transactions to purchase movie-ticket token and mint a movie-ticket token to user wallet
      parameters:
      - description: Purchase info
        in: body
        name: purchase_info
        required: true
        schema:
          $ref: '#/definitions/service.PurchaseInfo'
      - description: Base coin transfer session Token
        in: path
        name: baseCoinTransferToken
        required: true
        type: string
      - description: Base coin transfer session Token
        in: path
        name: movieTokenTransferToken
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction hashes has executed
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Commit a purchasing movie-ticket token
      tags:
      - ticket
  /ticket/purchase/extra:
    post:
      consumes:
      - application/json
      description: Request user to transfer movie-token used for discounting ticket price
      parameters:
      - description: Purchase info
        in: body
        name: purchase_info
        required: true
        schema:
          $ref: '#/definitions/service.PurchaseInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Session token and redirect url to transfer a token
          schema:
            $ref: '#/definitions/service.TransferRequestResult'
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Request user to purchase extra token
      tags:
      - ticket
...
swagger: "2.0"